Errores de escritura

+++ State.hpp
@@ -9,7 +9,7 @@ namespace Sonar{
 		
 		virtual void HandleInput( )=0;
 		
-		virtual void Update(int dt)=0;
+		virtual void Update(float dt)=0;
 		
 		virtual void Draw( float dt )=0;
